+++ libraries/LPS331/LPS331.h
@@ -62,7 +62,7 @@ class LPS331
     long readPressureRaw(void);
     float readTemperatureC(void);
     float readTemperatureF(void);
-    int readTemperatureRaw(void);
+    int16_t readTemperatureRaw(void);
 
     static float pressureToAltitudeMeters(float pressure_mbar, float altimeter_setting_mbar = 1013.25);
     static float pressureToAltitudeFeet(float pressure_inHg, float altimeter_setting_inHg = 29.9213);
